add:添加step具体类（暂无具体实现）

+++ User/Peripherals/Motor/Servo/Inc/servo.h
@@ -7,13 +7,13 @@
 #include"motor.h"
 typedef struct {
     //位置量 (num*360)/den 度
-    uint32_t      num;
-    uint32_t      den;
-    uint32_t      maxAngle;
+    int      num;
+    int      den;
+    int      maxAngle;
     //电源控制引脚 可有可无
     GPIO_TypeDef* powerCtrlPort;
     uint16_t      powerCtrlPin;
-    uint8_t       isPowerOn;
+    int       isPowerOn;
     //定时器设置pwm
     PWM*    pwm;
 
@@ -21,15 +21,15 @@ typedef struct {
 
 Servo* NewServo(GPIO_TypeDef*powerCtrlPort,uint16_t powerCtrlPin, PWM* pwm);
 void DeleteServo(Servo* servo);
-uint8_t ServoPowerOn(Servo* servo);
-uint8_t ServoPowerOnByAngle( Servo* servo,uint32_t num,uint32_t den);
-uint8_t ServoShutDown(Servo* servo);
-uint8_t ServoSetPosition( Servo* servo,uint32_t num,uint32_t den,uint32_t maxAngel);
-uint8_t ServoSetPositionByAngle( Servo* servo,float angle);
+int ServoPowerOn(Servo* servo);
+int ServoPowerOnByAngle( Servo* servo,int num,int den);
+int ServoShutDown(Servo* servo);
+int ServoSetPosition( Servo* servo,int num,int den,int maxAngel);
+int ServoSetPositionByAngle( Servo* servo,float angle);
 MotorInterface ServoInterface();
 // servo 不存在以下类型接口但其它的motor可能会存在
-// uint8_t ServoSetPositionByEncode(Servo* servo,int encode);
-// uint8_t ServoSetSpeed(Servo* servo,int pwmNum,int pwmDen);
-// uint8_t ServoSetSpeedByEncode(Servo* servo,int spEncode);
-// uint8_t ServoSetSpeedByAngel(Servo* servo,int num,int den);
+// int ServoSetPositionByEncode(Servo* servo,int encode);
+// int ServoSetSpeed(Servo* servo,int pwmNum,int pwmDen);
+// int ServoSetSpeedByEncode(Servo* servo,int spEncode);
+// int ServoSetSpeedByAngel(Servo* servo,int num,int den);
 #endif
