+++ User/Peripherals/Motor/Servo/Src/servo.c
@@ -24,7 +24,7 @@ void DeleteServo(Servo* servo){
     }
 }
 
-uint8_t ServoPowerOnByAngle( Servo* servo,uint32_t num,uint32_t den){
+int ServoPowerOnByAngle( Servo* servo,int num,int den){
     if(servo == NULL || servo->pwm == NULL){
         return 1;
     }
@@ -35,7 +35,7 @@ uint8_t ServoPowerOnByAngle( Servo* servo,uint32_t num,uint32_t den){
     PWMStart(servo->pwm);
     return ServoSetPosition(servo,num,den,0);
 }
-uint8_t ServoShutDown( Servo* servo){
+int ServoShutDown( Servo* servo){
     if(servo == NULL){
         return 1;
     }
@@ -48,7 +48,7 @@ uint8_t ServoShutDown( Servo* servo){
     servo->isPowerOn = 0;
     return 0;
 }
-uint8_t ServoSetPosition( Servo* servo,uint32_t num,uint32_t den,uint32_t maxAngel){
+int ServoSetPosition( Servo* servo,int num,int den,int maxAngel){
     if(servo == NULL || servo->pwm == NULL || den == 0||den<num){
         return 1;
     }
@@ -60,8 +60,8 @@ uint8_t ServoSetPosition( Servo* servo,uint32_t num,uint32_t den,uint32_t maxAng
     }
     servo->num=num;
     servo->den=den;
-    uint32_t xden = servo->den*40;
-    uint32_t xnum = 4*servo->num+den;
+    int xden = servo->den*40;
+    int xnum = 4*servo->num+den;
     
     
     if(PWMSetByRate(servo->pwm,xnum,xden) != 0){
@@ -69,7 +69,7 @@ uint8_t ServoSetPosition( Servo* servo,uint32_t num,uint32_t den,uint32_t maxAng
     }
     return 0;
 }
-uint8_t ServoSetPositionByAngle( Servo* servo,float angle){
+int ServoSetPositionByAngle( Servo* servo,float angle){
     if(servo == NULL){
         return 1;
     }
@@ -79,50 +79,50 @@ uint8_t ServoSetPositionByAngle( Servo* servo,float angle){
     if(angle > 360){
         angle = 360;
     }
-    uint32_t num = (uint32_t)(angle * 10);
-    uint32_t den = 3600;
+    int num = (int)(angle * 10);
+    int den = 3600;
     return ServoSetPosition(servo,num,den,0);
 }
 
 // 适配函数：powerOn
-static uint8_t ServoAdapterPowerOn(void* instance){
+static int ServoAdapterPowerOn(void* instance){
     if(instance == NULL) return 1;
     // Servo默认上电到0度位置
     return ServoPowerOnByAngle((Servo*)instance, 0, 3600);
 }
 
 // 适配函数：shutDown
-static uint8_t ServoAdapterShutDown(void* instance){
+static int ServoAdapterShutDown(void* instance){
     if(instance == NULL) return 1;
     return ServoShutDown((Servo*)instance);
 }
 
 // 适配函数：setPosition
-static uint8_t ServoAdapterSetPosition(void* instance, uint32_t numAngel, uint32_t denAngel, uint32_t maxAngel){
+static int ServoAdapterSetPosition(void* instance, int numAngel, int denAngel, int maxAngel){
     if(instance == NULL) return 1;
     return ServoSetPosition((Servo*)instance, numAngel, denAngel, maxAngel);
 }
 
 // 适配函数：setPositionByEncode (Servo不支持编码器)
-static uint8_t ServoAdapterSetPositionByEncode(void* instance, int encode){
+static int ServoAdapterSetPositionByEncode(void* instance, int encode){
     // Servo不支持编码器模式
     return 1;
 }
 
 // 适配函数：setSpeedByPwm (Servo不支持速度控制，但可以通过PWM设置位置)
-static uint8_t ServoAdapterSetSpeedByPwm(void* instance, int pwmNum, int pwmDen){
+static int ServoAdapterSetSpeedByPwm(void* instance, int pwmNum, int pwmDen){
     // Servo不支持速度控制模式
     return 1;
 }
 
 // 适配函数：setSpeedByAngel (Servo不支持速度控制)
-static uint8_t ServoAdapterSetSpeedByAngel(void* instance, int spNumAngel, int spDenAngel){
+static int ServoAdapterSetSpeedByAngel(void* instance, int spNumAngel, int spDenAngel){
     // Servo不支持速度控制模式
     return 1;
 }
 
 // 适配函数：setSpeedByEncode (Servo不支持编码器)
-static uint8_t ServoAdapterSetSpeedByEncode(void* instance, int encode){
+static int ServoAdapterSetSpeedByEncode(void* instance, int encode){
     // Servo不支持编码器模式
     return 1;
 }
