+++ User/Peripherals/Motor/motor.h
@@ -3,13 +3,13 @@
 #include <stdint.h>
 #include "cmsis_os.h"
 typedef struct{
-    uint8_t (*powerOn)(void*instance);
-    uint8_t (*shutDown)(void*instance);
-    uint8_t (*setPosition)(void*instance,uint32_t numAngel,uint32_t denAngel,uint32_t maxAngel);
-    uint8_t (*setPositionByEncode)(void*instance,int encode);
-    uint8_t (*setSpeedByPwm)(void*instance,int pwmNum,int pwmDen);
-    uint8_t (*setSpeedByAngel)(void*instance,int spNumAngel,int spDenAngel);
-    uint8_t (*setSpeedByEncode)(void*instance,int encode);
+    int (*powerOn)(void*instance);
+    int (*shutDown)(void*instance);
+    int (*setPosition)(void*instance,int numAngel,int denAngel,int maxAngel);
+    int (*setPositionByEncode)(void*instance,int encode);
+    int (*setSpeedByPwm)(void*instance,int pwmNum,int pwmDen);
+    int (*setSpeedByAngel)(void*instance,int spNumAngel,int spDenAngel);
+    int (*setSpeedByEncode)(void*instance,int encode);
 }MotorInterface;
 typedef struct {
     void*       instance;
